refact #204 : message 메타데이터 추가
Close #204
