+++ messages/ko.json
@@ -240,5 +240,14 @@
   "Chatbot": {
     "comment": "안녕하세요 아잇나우 챗봇입니다. 해외주식 관련해서 궁금하신 점이 있으면 저에게 물어보세요!",
     "send": "전송"
+  },
+  "Metadata": {
+    "app_name": "아잇나우",
+    "search": "발견",
+    "news": "뉴스",
+    "interest": "관심 종목",
+    "my": "마이 페이지",
+    "logout": "로그아웃",
+    "stock": "주식"
   }
 }
