+++ src/app/[locale]/loading.tsx
