--- src/app/layout.tsx
@@ -1,30 +1,0 @@
-import type { Metadata } from 'next';
-import './globals.css';
-import localFont from 'next/font/local';
-
-const pretendard = localFont({
-  src: '../assets/fonts/PretendardVariable.woff2',
-  display: 'swap',
-  weight: '45 920',
-  variable: '--font-pretendard',
-});
-
-export const metadata: Metadata = {
-  title: {
-    template: '%s | Eightnow',
-    default: 'Eightnow',
-  },
-  description: 'AI ANALYSIS REPORT PLATFORM',
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="ko" className={`${pretendard.variable}`}>
-      <body className={`${pretendard.className}`}>{children}</body>
-    </html>
-  );
-}
